requirements.txt
requests==2.32.3
urllib3==2.2.3
certifi==2025.1.31
python-dotenv==1.0.1
pytz==2024.2
gspread==6.1.4
google-auth==2.34.0